+++ src/main/scala/edu/ie3/simona/model/grid/TransformerModel.scala
@@ -19,10 +19,8 @@ import edu.ie3.util.quantities.PowerSystemUnits.*
 import edu.ie3.util.scala.OperationInterval
 import edu.ie3.util.scala.quantities.ApparentPower
 import edu.ie3.util.scala.quantities.QuantityConversionUtils.{
-  OhmToSimona,
-  PowerConversionSimona,
-  SiemensToSimona,
-  VoltageToSimona,
+  toApparent,
+  toSquants,
 }
 import squants.Each